+++ testing/matrix-sdk-integration-testing/src/tests/e2ee.rs
@@ -5,11 +5,14 @@ use assert_matches::assert_matches;
 use assign::assign;
 use matrix_sdk::{
     crypto::{format_emojis, SasState},
-    encryption::{verification::VerificationRequestState, LocalTrust},
+    encryption::{
+        verification::{QrVerificationData, QrVerificationState, VerificationRequestState},
+        LocalTrust,
+    },
     ruma::{
         api::client::room::create_room::v3::Request as CreateRoomRequest,
         events::{
-            key::verification::request::ToDeviceKeyVerificationRequestEvent,
+            key::verification::{request::ToDeviceKeyVerificationRequestEvent, VerificationMethod},
             room::message::{
                 MessageType, OriginalSyncRoomMessageEvent, RoomMessageEventContent,
                 SyncRoomMessageEvent,
@@ -23,7 +26,7 @@ use tracing::warn;
 use crate::helpers::{SyncTokenAwareClient, TestClientBuilder};
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn test_mutual_verification() -> Result<()> {
+async fn test_mutual_sas_verification() -> Result<()> {
     let alice = SyncTokenAwareClient::new(
         TestClientBuilder::new("alice")
             .randomize_username()
@@ -52,11 +55,12 @@ async fn test_mutual_verification() -> Result<()> {
 
     let alice_room = alice.create_room(request).await?;
     alice_room.enable_encryption().await?;
+    let room_id = alice_room.room_id();
 
     warn!("alice has created and enabled encryption in the room");
 
     bob.sync_once().await?;
-    bob.get_room(alice_room.room_id()).unwrap().join().await?;
+    bob.get_room(room_id).unwrap().join().await?;
 
     alice.sync_once().await?;
 
@@ -104,6 +108,7 @@ async fn test_mutual_verification() -> Result<()> {
 
     assert!(!alice_verification_request.is_passive());
     assert!(alice_verification_request.we_started());
+    assert_eq!(alice_verification_request.room_id(), Some(room_id));
 
     warn!("alice has started verification");
 
@@ -115,6 +120,8 @@ async fn test_mutual_verification() -> Result<()> {
         .expect("bob received a verification request");
 
     warn!("bob has received the verification request");
+
+    assert_eq!(bob_verification_request.room_id(), Some(room_id));
 
     assert!(!bob_verification_request.is_done());
 
@@ -151,6 +158,7 @@ async fn test_mutual_verification() -> Result<()> {
 
     assert_eq!(alice_sas.own_user_id(), alice.user_id().unwrap());
     assert_eq!(alice_sas.other_user_id(), bob.user_id().unwrap());
+    assert_eq!(alice_sas.room_id(), Some(room_id));
 
     assert!(alice_sas.started_from_request());
     assert!(!alice_sas.is_self_verification());
@@ -173,6 +181,7 @@ async fn test_mutual_verification() -> Result<()> {
     assert!(bob_verification.cancel_info().is_none());
     assert_eq!(bob_verification.own_user_id(), bob.user_id().unwrap());
     assert_eq!(bob_verification.other_user_id(), alice.user_id().unwrap());
+    assert_eq!(bob_verification.room_id(), Some(room_id));
 
     let bob_sas = bob_verification.sas().unwrap();
 
@@ -272,6 +281,245 @@ async fn test_mutual_verification() -> Result<()> {
     assert!(bob_alice_device.is_verified());
     assert_eq!(bob_alice_device.local_trust_state(), LocalTrust::Verified);
     assert!(bob_alice_device.is_locally_trusted());
+    assert!(!bob_alice_device.is_blacklisted());
+
+    Ok(())
+}
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
+async fn test_mutual_qrcode_verification() -> Result<()> {
+    let alice = SyncTokenAwareClient::new(
+        TestClientBuilder::new("alice")
+            .randomize_username()
+            .use_sqlite()
+            .bootstrap_cross_signing()
+            .build()
+            .await?,
+    );
+    let bob = SyncTokenAwareClient::new(
+        TestClientBuilder::new("bob")
+            .randomize_username()
+            .use_sqlite()
+            .bootstrap_cross_signing()
+            .build()
+            .await?,
+    );
+
+    warn!("alice's device: {}", alice.device_id().unwrap());
+    warn!("bob's device: {}", bob.device_id().unwrap());
+
+    let invite = vec![bob.user_id().unwrap().to_owned()];
+    let request = assign!(CreateRoomRequest::new(), {
+        invite,
+        is_direct: true,
+    });
+
+    let alice_room = alice.create_room(request).await?;
+    alice_room.enable_encryption().await?;
+    let room_id = alice_room.room_id();
+
+    warn!("alice has created and enabled encryption in the room");
+
+    bob.sync_once().await?;
+    bob.get_room(room_id).unwrap().join().await?;
+
+    alice.sync_once().await?;
+
+    warn!("alice and bob are both aware of each other in the e2ee room");
+
+    // Bob adds the verification listeners.
+    let bob_verification_request = Arc::new(Mutex::new(None));
+    {
+        let bvr = bob_verification_request.clone();
+        bob.add_event_handler(
+            |ev: ToDeviceKeyVerificationRequestEvent, client: Client| async move {
+                let request = client
+                    .encryption()
+                    .get_verification_request(&ev.sender, &ev.content.transaction_id)
+                    .await
+                    .expect("Request object wasn't created");
+                *bvr.lock().unwrap() = Some(request);
+            },
+        );
+
+        let bvr = bob_verification_request.clone();
+        bob.add_event_handler(|ev: OriginalSyncRoomMessageEvent, client: Client| async move {
+            if let MessageType::VerificationRequest(_) = &ev.content.msgtype {
+                let request = client
+                    .encryption()
+                    .get_verification_request(&ev.sender, &ev.event_id)
+                    .await
+                    .expect("Request object wasn't created");
+                *bvr.lock().unwrap() = Some(request);
+            }
+        });
+    }
+
+    warn!("bob has set up verification listeners");
+
+    let alice_bob_identity = alice
+        .encryption()
+        .get_user_identity(bob.user_id().unwrap())
+        .await?
+        .expect("alice knows bob's identity");
+
+    warn!("alice has found bob's identity");
+
+    let alice_verification_request = alice_bob_identity
+        .request_verification_with_methods(vec![
+            VerificationMethod::SasV1,
+            VerificationMethod::QrCodeScanV1,
+            VerificationMethod::QrCodeShowV1,
+            VerificationMethod::ReciprocateV1,
+        ])
+        .await?;
+
+    assert!(!alice_verification_request.is_passive());
+    assert!(alice_verification_request.we_started());
+    assert_eq!(alice_verification_request.room_id(), Some(room_id));
+
+    warn!("alice has started verification");
+
+    bob.sync_once().await?;
+    let bob_verification_request = bob_verification_request
+        .lock()
+        .unwrap()
+        .take()
+        .expect("bob received a verification request");
+
+    warn!("bob has received the verification request");
+
+    assert_eq!(bob_verification_request.room_id(), Some(room_id));
+
+    assert!(!bob_verification_request.is_done());
+
+    assert!(!bob_verification_request.is_cancelled());
+    assert!(bob_verification_request.cancel_info().is_none());
+
+    assert!(!bob_verification_request.is_ready());
+    assert!(!bob_verification_request.is_passive());
+    assert!(!bob_verification_request.we_started());
+
+    assert_eq!(bob_verification_request.own_user_id(), bob.user_id().unwrap());
+    assert_eq!(bob_verification_request.other_user_id(), alice.user_id().unwrap());
+    assert!(!bob_verification_request.is_self_verification());
+
+    assert_matches!(bob_verification_request.state(), VerificationRequestState::Requested { .. });
+    let _flow_id = bob_verification_request.flow_id();
+
+    // Bob notifies Alice he accepts the verification process.
+    bob_verification_request.accept().await.unwrap();
+    assert_matches!(bob_verification_request.state(), VerificationRequestState::Ready { .. });
+    assert!(bob_verification_request.is_ready());
+
+    warn!("bob has accepted the verification request");
+
+    // Alice receives the accept, and moves to the ready state.
+    assert_matches!(alice_verification_request.state(), VerificationRequestState::Created { .. });
+    alice.sync_once().await.unwrap();
+    assert_matches!(alice_verification_request.state(), VerificationRequestState::Ready { .. });
+
+    let bob_qr =
+        bob_verification_request.generate_qr_code().await?.expect("must have a qr verification");
+
+    assert_matches!(bob_qr.state(), QrVerificationState::Started);
+
+    warn!("bob has generated the QR code");
+
+    let bob_verification = assert_matches!(bob_verification_request.state(), VerificationRequestState::Transitioned { verification } => verification);
+
+    assert!(!bob_verification.is_done());
+    assert!(!bob_verification.is_cancelled());
+    assert!(!bob_verification.is_self_verification());
+    assert!(!bob_verification.we_started());
+    assert!(bob_verification.cancel_info().is_none());
+    assert_eq!(bob_verification.own_user_id(), bob.user_id().unwrap());
+    assert_eq!(bob_verification.other_user_id(), alice.user_id().unwrap());
+    assert_eq!(bob_verification.room_id(), Some(room_id));
+
+    let qr_code_bytes = bob_qr.to_bytes()?;
+    let alice_qr = alice_verification_request
+        .scan_qr_code(QrVerificationData::from_bytes(qr_code_bytes)?)
+        .await?
+        .expect("must have a qr verification");
+
+    warn!("alice has scanned the QR code");
+
+    assert_eq!(alice_qr.own_user_id(), alice.user_id().unwrap());
+    assert_eq!(alice_qr.other_user_id(), bob.user_id().unwrap());
+    assert_eq!(alice_qr.room_id(), Some(room_id));
+
+    assert!(!alice_qr.is_self_verification());
+    assert!(alice_qr.we_started());
+    assert!(!alice_qr.is_done());
+    assert!(!alice_qr.is_cancelled());
+    assert!(alice_qr.cancel_info().is_none());
+
+    bob.sync_once().await?;
+
+    assert_matches!(alice_qr.state(), QrVerificationState::Reciprocated);
+    assert_matches!(bob_qr.state(), QrVerificationState::Scanned);
+
+    bob_qr.confirm().await?;
+
+    warn!("bob has confirmed the QR code scanning");
+
+    alice.sync_once().await?;
+
+    assert_matches!(bob_qr.state(), QrVerificationState::Confirmed);
+    assert_matches!(alice_qr.state(), QrVerificationState::Done { .. });
+
+    // Crypto dancing.
+    alice.sync_once().await?;
+    bob.sync_once().await?;
+    assert_matches!(bob_qr.state(), QrVerificationState::Done { .. });
+    assert_matches!(alice_qr.state(), QrVerificationState::Done { .. });
+
+    // Wait for remote echos for verification status requests.
+    alice.sync_once().await?;
+    bob.sync_once().await?;
+
+    assert!(!bob_verification_request.is_cancelled());
+    assert!(!alice_verification_request.is_cancelled());
+
+    assert!(bob_verification_request.is_done());
+    assert!(alice_verification_request.is_done());
+    assert!(bob_qr.is_done());
+    assert!(alice_qr.is_done());
+
+    // Both users appear as verified to each other.
+    let alice_bob_ident =
+        alice.encryption().get_user_identity(bob.user_id().unwrap()).await?.unwrap();
+    assert!(alice_bob_ident.is_verified());
+
+    let bob_alice_ident =
+        bob.encryption().get_user_identity(alice.user_id().unwrap()).await?.unwrap();
+    assert!(bob_alice_ident.is_verified());
+
+    // Both user devices appear as verified to the other user.
+    let alice_bob_device = alice_qr.other_device();
+    assert_eq!(alice_bob_device.user_id(), bob.user_id().unwrap());
+    assert_eq!(alice_bob_device.device_id(), bob.device_id().unwrap());
+    assert_eq!(alice_bob_device.local_trust_state(), LocalTrust::Unset);
+
+    let alice_bob_device = alice
+        .encryption()
+        .get_device(bob.user_id().unwrap(), bob.device_id().unwrap())
+        .await?
+        .unwrap();
+    assert!(alice_bob_device.is_verified());
+    assert!(!alice_bob_device.is_blacklisted());
+
+    let bob_alice_device = bob_qr.other_device();
+    assert_eq!(bob_alice_device.user_id(), alice.user_id().unwrap());
+    assert_eq!(bob_alice_device.device_id(), alice.device_id().unwrap());
+
+    let bob_alice_device = bob
+        .encryption()
+        .get_device(alice.user_id().unwrap(), alice.device_id().unwrap())
+        .await?
+        .unwrap();
+    assert!(bob_alice_device.is_verified());
     assert!(!bob_alice_device.is_blacklisted());
 
     Ok(())